+++ examples/benchmark.py
@@ -5,10 +5,11 @@ Usage (kwargs optional): python3 benchmark.py cycles=10000
 
 Created on 5 Nov 2021
 
-:author: semuadmin
-:copyright: SEMU Consulting © 2021
+:author: semuadmin (Steve Smith)
+:copyright: semuadmin © 2021
 :license: BSD 3-Clause
 """
+
 # pylint: disable=line-too-long
 
 from sys import argv