+++ src/pynmeagps/nmeamessage.py
@@ -3,8 +3,8 @@ Main NMEA GNSS/GPS Message Protocol Class.
 
 Created on 04 Mar 2021
 
-:author: semuadmin
-:copyright: SEMU Consulting © 2021
+:author: semuadmin (Steve Smith)
+:copyright: semuadmin © 2021
 :license: BSD 3-Clause
 """