+++ src/pynmeagps/nmeatypes_get.py
@@ -24,7 +24,7 @@ Created on 4 Mar Sep 2021
 While the NMEA 0183 © protocol is proprietary, the information here
 has been collated from public domain sources.
 
-:author: semuadmin
+:author: semuadmin (Steve Smith)
 """
 
 # pylint: disable=too-many-lines, duplicate-code