+++ src/pynmeagps/nmeatypes_poll.py
@@ -12,7 +12,7 @@ Created on 4 Mar Sep 2021
 While the NMEA 0183 © protocol is proprietary, the information here
 has been collated from public domain sources.
 
-:author: semuadmin
+:author: semuadmin (Steve Smith)
 """
 
 from pynmeagps.nmeatypes_core import ST