+++ tests/test_parse.py
@@ -6,7 +6,7 @@ Created on 4 Mar 2021
 
 *** NB: must be saved in UTF-8 format ***
 
-:author: semuadmin
+:author: semuadmin (Steve Smith)
 """
 
 import unittest